+++ src/images/image_crate.rs
@@ -4,7 +4,7 @@ use bytemuck::cast_slice_mut;
 use image::DynamicImage;
 
 use crate::image_view::try_pixel_type;
-use crate::images::{TypedImage, TypedImageMut};
+use crate::images::{TypedImage, TypedImageRef};
 use crate::pixels::InnerPixel;
 use crate::{ImageView, ImageViewMut, IntoImageView, IntoImageViewMut, PixelType};
 
@@ -34,8 +34,12 @@ impl IntoImageView for DynamicImage {
     fn image_view<P: InnerPixel>(&self) -> Option<impl ImageView<Pixel = P>> {
         if let Ok(pixel_type) = try_pixel_type(self) {
             if P::pixel_type() == pixel_type {
-                return TypedImage::<P>::from_buffer(self.width(), self.height(), self.as_bytes())
-                    .ok();
+                return TypedImageRef::<P>::from_buffer(
+                    self.width(),
+                    self.height(),
+                    self.as_bytes(),
+                )
+                .ok();
             }
         }
         None
@@ -46,7 +50,7 @@ impl IntoImageViewMut for DynamicImage {
     fn image_view_mut<P: InnerPixel>(&mut self) -> Option<impl ImageViewMut<Pixel = P>> {
         if let Ok(pixel_type) = try_pixel_type(self) {
             if P::pixel_type() == pixel_type {
-                return TypedImageMut::<P>::from_buffer(
+                return TypedImage::<P>::from_buffer(
                     self.width(),
                     self.height(),
                     image_as_bytes_mut(self),