+++ tests/alpha_tests.rs
@@ -1,4 +1,4 @@
-use fast_image_resize::images::{Image, TypedImage, TypedImageMut};
+use fast_image_resize::images::{Image, TypedImage, TypedImageRef};
 use fast_image_resize::{CpuExtensions, MulDiv, PixelTrait};
 use testing::{cpu_ext_into_str, PixelTestingExt};
 
@@ -83,8 +83,8 @@ fn mul_div_alpha_test<P: PixelTrait>(
         .collect();
     let mut dst_pixels = src_pixels.clone();
 
-    let src_image = TypedImage::from_pixels(width, height, &src_pixels).unwrap();
-    let mut dst_image = TypedImageMut::from_pixels(width, height, &mut dst_pixels).unwrap();
+    let src_image = TypedImageRef::new(width, height, &src_pixels).unwrap();
+    let mut dst_image = TypedImage::from_pixels(width, height, &mut dst_pixels).unwrap();
 
     let mut alpha_mul_div: MulDiv = Default::default();
     unsafe {
@@ -125,7 +125,7 @@ fn mul_div_alpha_test<P: PixelTrait>(
 
     // Inplace
     let mut src_pixels_clone = src_pixels.clone();
-    let mut image = TypedImageMut::from_pixels(width, height, &mut src_pixels_clone).unwrap();
+    let mut image = TypedImage::from_pixels(width, height, &mut src_pixels_clone).unwrap();
 
     match oper {
         Oper::Mul => alpha_mul_div